agui_middleware_py/requirements.txt
# AGUI Middleware Python Requirements
# Core AGUI SDK
ag-ui>=0.1.0

# ISEK Integration
isek
a2a

# FastAPI for HTTP server
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Async support
asyncio
aiohttp
nest-asyncio

# Environment and configuration
python-dotenv
pyyaml

# Logging and monitoring
structlog
rich

# Testing
pytest
pytest-asyncio
httpx

# Development
black
flake8